Application: compile shaders in an application base class
Since OpenGL requires a context to be set (this means that window should be set as well)
to perform any of its operations (including shaders compilation), the decision was made
to compile shaders in a some generic component, where a context is set at the moment of execution.
For now, Application base class fits those requirements the best way possible.